zcash_client_backend: Migrate `send_single_step_proposed_transfer` test from `zcash_client_sqlite`
