+++ zcash_client_sqlite/src/testing/pool.rs
@@ -12,21 +12,6 @@ use incrementalmerkletree::frontier::Frontier;
 use rusqlite::params;
 use secrecy::Secret;
 
-use zcash_primitives::{
-    block::BlockHash,
-    consensus::{BranchId, NetworkUpgrade, Parameters},
-    legacy::TransparentAddress,
-    memo::{Memo, MemoBytes},
-    transaction::{
-        components::amount::NonNegativeAmount,
-        fees::{
-            fixed::FeeRule as FixedFeeRule, zip317::FeeError as Zip317FeeError, StandardFeeRule,
-        },
-        Transaction,
-    },
-    zip32::Scope,
-};
-
 use zcash_client_backend::{
     address::Address,
     data_api::{
@@ -37,19 +22,29 @@ use zcash_client_backend::{
             input_selector, pool::ShieldedPoolTester, sapling::SaplingPoolTester, AddressType,
             FakeCompactOutput, InitialChainState, TestBuilder, TestState,
         },
-        wallet::{
-            decrypt_and_store_transaction,
-            input_selection::{GreedyInputSelector, GreedyInputSelectorError},
-        },
+        wallet::input_selection::{GreedyInputSelector, GreedyInputSelectorError},
         Account as _, AccountBirthday, Ratio, WalletRead, WalletWrite,
     },
-    decrypt_transaction,
     fees::{fixed, standard, DustOutputPolicy},
     keys::UnifiedSpendingKey,
     scanning::ScanError,
     wallet::{Note, OvkPolicy},
-    zip321::{self, Payment, TransactionRequest},
+    zip321::{Payment, TransactionRequest},
 };
+use zcash_primitives::{
+    block::BlockHash,
+    consensus::{BranchId, NetworkUpgrade, Parameters},
+    legacy::TransparentAddress,
+    transaction::{
+        components::amount::NonNegativeAmount,
+        fees::{
+            fixed::FeeRule as FixedFeeRule, zip317::FeeError as Zip317FeeError, StandardFeeRule,
+        },
+        Transaction,
+    },
+    zip32::Scope,
+};
+use zcash_protocol::memo::MemoBytes;
 
 use crate::{
     error::SqliteClientError,
@@ -58,7 +53,7 @@ use crate::{
         BlockCache,
     },
     wallet::{commitment_tree, parse_scope, truncate_to_height},
-    NoteId, ReceivedNoteId, SAPLING_TABLES_PREFIX,
+    ReceivedNoteId, SAPLING_TABLES_PREFIX,
 };
 
 #[cfg(feature = "transparent-inputs")]
@@ -69,6 +64,7 @@ use {
         components::{OutPoint, TxOut},
         fees::zip317,
     },
+    zcash_protocol::memo::Memo,
 };
 
 #[cfg(feature = "orchard")]
@@ -92,156 +88,10 @@ impl ShieldedPoolPersistence for OrchardPoolTester {
 }
 
 pub(crate) fn send_single_step_proposed_transfer<T: ShieldedPoolTester>() {
-    let mut st = TestBuilder::new()
-        .with_data_store_factory(TestDbFactory)
-        .with_block_cache(BlockCache::new())
-        .with_account_from_sapling_activation(BlockHash([0; 32]))
-        .build();
-
-    let account = st.test_account().cloned().unwrap();
-    let dfvk = T::test_account_fvk(&st);
-
-    // Add funds to the wallet in a single note
-    let value = NonNegativeAmount::const_from_u64(60000);
-    let (h, _, _) = st.generate_next_block(&dfvk, AddressType::DefaultExternal, value);
-    st.scan_cached_blocks(h, 1);
-
-    // Spendable balance matches total balance
-    assert_eq!(st.get_total_balance(account.id()), value);
-    assert_eq!(st.get_spendable_balance(account.id(), 1), value);
-
-    assert_eq!(
-        st.wallet()
-            .block_max_scanned()
-            .unwrap()
-            .unwrap()
-            .block_height(),
-        h
-    );
-
-    let to_extsk = T::sk(&[0xf5; 32]);
-    let to: Address = T::sk_default_address(&to_extsk);
-    let request = zip321::TransactionRequest::new(vec![Payment::without_memo(
-        to.to_zcash_address(st.network()),
-        NonNegativeAmount::const_from_u64(10000),
-    )])
-    .unwrap();
-
-    let fee_rule = StandardFeeRule::Zip317;
-
-    let change_memo = "Test change memo".parse::<Memo>().unwrap();
-    let change_strategy = standard::SingleOutputChangeStrategy::new(
-        fee_rule,
-        Some(change_memo.clone().into()),
-        T::SHIELDED_PROTOCOL,
-    );
-    let input_selector = &GreedyInputSelector::new(change_strategy, DustOutputPolicy::default());
-
-    let proposal = st
-        .propose_transfer(
-            account.id(),
-            input_selector,
-            request,
-            NonZeroU32::new(1).unwrap(),
-        )
-        .unwrap();
-
-    let create_proposed_result = st.create_proposed_transactions::<Infallible, _>(
-        account.usk(),
-        OvkPolicy::Sender,
-        &proposal,
-    );
-    assert_matches!(&create_proposed_result, Ok(txids) if txids.len() == 1);
-
-    let sent_tx_id = create_proposed_result.unwrap()[0];
-
-    // Verify that the sent transaction was stored and that we can decrypt the memos
-    let tx = st
-        .wallet()
-        .get_transaction(sent_tx_id)
-        .unwrap()
-        .expect("Created transaction was stored.");
-    let ufvks = [(account.id(), account.usk().to_unified_full_viewing_key())]
-        .into_iter()
-        .collect();
-    let d_tx = decrypt_transaction(st.network(), h + 1, &tx, &ufvks);
-    assert_eq!(T::decrypted_pool_outputs_count(&d_tx), 2);
-
-    let mut found_tx_change_memo = false;
-    let mut found_tx_empty_memo = false;
-    T::with_decrypted_pool_memos(&d_tx, |memo| {
-        if Memo::try_from(memo).unwrap() == change_memo {
-            found_tx_change_memo = true
-        }
-        if Memo::try_from(memo).unwrap() == Memo::Empty {
-            found_tx_empty_memo = true
-        }
-    });
-    assert!(found_tx_change_memo);
-    assert!(found_tx_empty_memo);
-
-    // Verify that the stored sent notes match what we're expecting
-    let sent_note_ids = {
-        let mut stmt_sent_notes = st
-            .wallet()
-            .conn()
-            .prepare(
-                "SELECT output_index
-                FROM sent_notes
-                JOIN transactions ON transactions.id_tx = sent_notes.tx
-                WHERE transactions.txid = ?",
-            )
-            .unwrap();
-
-        stmt_sent_notes
-            .query(rusqlite::params![sent_tx_id.as_ref()])
-            .unwrap()
-            .mapped(|row| Ok(NoteId::new(sent_tx_id, T::SHIELDED_PROTOCOL, row.get(0)?)))
-            .collect::<Result<Vec<_>, _>>()
-            .unwrap()
-    };
-
-    assert_eq!(sent_note_ids.len(), 2);
-
-    // The sent memo should be the empty memo for the sent output, and the
-    // change output's memo should be as specified.
-    let mut found_sent_change_memo = false;
-    let mut found_sent_empty_memo = false;
-    for sent_note_id in sent_note_ids {
-        match st
-            .wallet()
-            .get_memo(sent_note_id)
-            .expect("Note id is valid")
-            .as_ref()
-        {
-            Some(m) if m == &change_memo => {
-                found_sent_change_memo = true;
-            }
-            Some(m) if m == &Memo::Empty => {
-                found_sent_empty_memo = true;
-            }
-            Some(other) => panic!("Unexpected memo value: {:?}", other),
-            None => panic!("Memo should not be stored as NULL"),
-        }
-    }
-    assert!(found_sent_change_memo);
-    assert!(found_sent_empty_memo);
-
-    // Check that querying for a nonexistent sent note returns None
-    assert_matches!(
-        st.wallet()
-            .get_memo(NoteId::new(sent_tx_id, T::SHIELDED_PROTOCOL, 12345)),
-        Ok(None)
-    );
-
-    let tx_history = st.wallet().get_tx_history().unwrap();
-    assert_eq!(tx_history.len(), 2);
-
-    let network = *st.network();
-    assert_matches!(
-        decrypt_and_store_transaction(&network, st.wallet_mut(), &tx, None),
-        Ok(_)
-    );
+    zcash_client_backend::data_api::testing::pool::send_single_step_proposed_transfer::<T>(
+        TestDbFactory,
+        BlockCache::new(),
+    )
 }
 
 #[cfg(feature = "transparent-inputs")]
@@ -1755,7 +1605,7 @@ pub(crate) fn pool_crossing_required<P0: ShieldedPoolTester, P1: ShieldedPoolTes
     assert_eq!(st.get_spendable_balance(account.id(), 1), initial_balance);
 
     let transfer_amount = NonNegativeAmount::const_from_u64(200000);
-    let p0_to_p1 = zip321::TransactionRequest::new(vec![Payment::without_memo(
+    let p0_to_p1 = TransactionRequest::new(vec![Payment::without_memo(
         p1_to.to_zcash_address(st.network()),
         transfer_amount,
     )])
@@ -1843,7 +1693,7 @@ pub(crate) fn fully_funded_fully_private<P0: ShieldedPoolTester, P1: ShieldedPoo
     assert_eq!(st.get_spendable_balance(account.id(), 1), initial_balance);
 
     let transfer_amount = NonNegativeAmount::const_from_u64(200000);
-    let p0_to_p1 = zip321::TransactionRequest::new(vec![Payment::without_memo(
+    let p0_to_p1 = TransactionRequest::new(vec![Payment::without_memo(
         p1_to.to_zcash_address(st.network()),
         transfer_amount,
     )])
@@ -1931,7 +1781,7 @@ pub(crate) fn fully_funded_send_to_t<P0: ShieldedPoolTester, P1: ShieldedPoolTes
     assert_eq!(st.get_spendable_balance(account.id(), 1), initial_balance);
 
     let transfer_amount = NonNegativeAmount::const_from_u64(200000);
-    let p0_to_p1 = zip321::TransactionRequest::new(vec![Payment::without_memo(
+    let p0_to_p1 = TransactionRequest::new(vec![Payment::without_memo(
         Address::Transparent(p1_to).to_zcash_address(st.network()),
         transfer_amount,
     )])
@@ -2043,7 +1893,7 @@ pub(crate) fn multi_pool_checkpoint<P0: ShieldedPoolTester, P1: ShieldedPoolTest
 
     // First, send funds just to P0
     let transfer_amount = NonNegativeAmount::const_from_u64(200000);
-    let p0_transfer = zip321::TransactionRequest::new(vec![Payment::without_memo(
+    let p0_transfer = TransactionRequest::new(vec![Payment::without_memo(
         P0::random_address(st.rng_mut()).to_zcash_address(st.network()),
         transfer_amount,
     )])
@@ -2068,7 +1918,7 @@ pub(crate) fn multi_pool_checkpoint<P0: ShieldedPoolTester, P1: ShieldedPoolTest
     assert_eq!(st.get_pending_change(acct_id, 1), expected_change);
 
     // In the next block, send funds to both P0 and P1
-    let both_transfer = zip321::TransactionRequest::new(vec![
+    let both_transfer = TransactionRequest::new(vec![
         Payment::without_memo(
             P0::random_address(st.rng_mut()).to_zcash_address(st.network()),
             transfer_amount,
